ml/requirements.txt
accelerate>=0.33.0
bitsandbytes>=0.45.0
datasets>=2.20.0
google-generativeai>=0.8.0
peft>=0.12.0
pypdf>=5.0.0
PyMuPDF>=1.24.0
python-docx>=1.1.2
requests>=2.32.3
sentencepiece>=0.2.0
torch>=2.3.0
transformers>=4.51.0
trl>=0.10.1
